+++ examples/wpf-controls/WpfControls/App.xaml.cs
@@ -1,0 +1,8 @@
+using System.Windows;
+
+namespace WpfControls
+{
+    public partial class App : Application
+    {
+    }
+}